Use Traits to simplify Sccp_Manager.class
Move helper functions to trait
Move Ajax to trait (WIP - next step to create subclass to simplify initialisation)
Rename Directories to eliminate punctuation
Add new directory for traits
