+++ SccpManClasses/amInterfaceClasses/Event.class.php
