+++ SccpManClasses/amInterfaceClasses/Message.class.php
