+++ SccpManClasses/amInterfaceClasses/Response.class.php
