+++ SccpManClasses/dbinterface.class.php
