+++ SccpManClasses/extconfigs.class.php
