+++ SccpManClasses/sipconfigs.class.php
