+++ SccpManClasses/xmlinterface.class.php
